Add files via upload
The files uploaded contain the python notebook, and the presentation based on the code with all the results. Check the notebook for the code, the PPT for the results and from the dataset check out the link shared below.  
This project is based on a dataset given on Kaggle: https://www.kaggle.com/mlg-ulb/creditcardfraud. The idea is to use a classifier to identify fraudulent an genuine transactions (which are 0.24%). We did Exploratory Data Analysis on key variables like Amount and Time. For Data transformation we used Synthetic Minority Oversampling technique considering there is heavy class imbalances. We use three different models: Logistic Regression, XGBoost and Multi-Layer Perceptron Classifier. We use 5 fold cross validation to check fitness We test the model on a fraud only dataset and on the top 10 fraud transactions For Logistic and XGBoost we got the features that are most important.